% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the file name.}
}
\value{
A data frame with data from the csv file or an error if the
  file does not exists.
}
\description{
This function reads in a .csv file from the \strong{US
National Highway Traffic Safety Administration's} \emph{Fatality Analysis
Reporting System} (FARS).
}
\details{
For more information, visit:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}
\seealso{
\link{make_filename}
}
