% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load multiple years of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of four-digit years of the FARS data to load.  Currently only data from 2013-2015 are available.
\code{years} must be coerceable to an integer, or else an error will be generated.}
}
\value{
This function returns a list of dplyr tibbles containing FARS data for each year in \code{years}.
   The tibbles contain only the month and year of each accident, so the list is mostly of use as an
   input into other functions.   If any element in \code{years} does not correspond to a valid FARS data
   file, a warning is generated and the corresponding component of the returned list is NULL
}
\description{
This function takes a vector of \code{years}, and returns a list of dplyr tibbles containting the
FARS data for that year.  The tibble contians only the month of the accident an the year in the
format of the \code{years} vector.  This is primarily a helper function for
\code{\link{fars_summarize_years}} as much of the data contained in the files is lost.
}
\details{
Any component of \code{years} which does not correspond to a valid filename will produce a warning, but a list
will still be returned, with NULL in position for each invalid component.

The FARS data can be found at the National Highway Traffic Safety Association
     \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
     2013-2015 data is distributed with this package.  The location of the files can be found using
     \code{system.file('extdata',package='FARScoursera')}
     These files need to be in your working directory for this function to work correctly.
}
\examples{
fars_read_years(2013:2015))
fars_read_years(c('2013','2014','2015'))
fars_read_years(2012:2016) # shows effect of invalid years in input

}
\seealso{
\code{\link{fars_read}}
}

