% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthly count of accidents by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of four-digit years of the FARS data to load.  Currently only data from 2013-2015 are available.
\code{years} must be coerceable to an integer, or else an error will be generated.}
}
\value{
This function returns a dplyr tibble with the month in rows and the year in columns.
    The entries are the total number of accidents in that year / month.  If \code{years} is a mix
    of valid and invalid years, the tibble is still returned by with warnings.  If \code{years}
    does not contain any valid years, then an error is thrown
}
\description{
This function produces a monthly count of the number of accidents in the FARS database for each year
given in \code{years}.
}
\details{
The FARS data can be found at the National Highway Traffic Safety Association
     \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
     2013-2015 data is distributed with this package.  The location of the files can be found using
     \code{system.file('extdata',package='FARScoursera')}
     These files need to be in your working directory for this function to work correctly.
}
\examples{
\dontrun{fars_summarize_years(c('2013','2014'))}
\dontrun{fars_summarize_years(2013:2014)}

}

