% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{FARSfunctions}
\alias{FARSfunctions}
\alias{fars_read}
\alias{make_filename}
\alias{fars_read_years}
\alias{fars_summarize_years}
\alias{fars_map_state}
\title{This set of functions are used to analyze the Fatality Analysis Reporting datasets.}
\usage{
fars_read(filename)

make_filename(year)

fars_read_years(years)

fars_summarize_years(years)

fars_map_state(state.num, year)
}
\arguments{
\item{filename}{A string giving the file name of dataset}

\item{year}{A string which defines a year, 2013, 2014, or 2015}

\item{years}{A svector containing strings which contains 1 or more years.}

\item{state.num}{A number which represents a state}
}
\value{
See description.
}
\description{
\itemize{} 	\code{fars_read()} takes a string (file name), and returns a data frame containing the corresponding data.
If the filename is not correct it will results in an error.
\itemize{} 	\code{make_file()} takes a string (year), and returns a string which is the file name of
the data for that particular year. If the year is not 2013-2015, there will be an error.
\itemize{} 	\code{fars_read_years()} does the same as fars_read, but takes a vector of strings (years)
and returns a list containing dataframes.If the years are not 2013-2015, there will be an error.
\itemize{} \code{fars_summarize_years()} takes a vector of strings (years) and returns a tibble containing
the number of fatal injuries by year and month.If the years are not 2013-2015, there will be an error.
\itemize{} \code{fars_map_state()} takes a state number and year, and returns a plot showing all traffic fatalities.
If the years are not 2013-2015, or the number doesn't correspond to a state, there will be an error.
}
\examples{
head(fars_read("accident_2013.csv.bz2"))
make_filename("2014")
fars_read_years(c("2014","2015"))
fars_summarize_years(c("2014","2015"))
fars_map_state(state.num = "10", year = "2015")
}
