% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a vaild FARS filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A four-digit year, in a format coerceable into an integer.  Currently, only files
covering 2013-2013 are valid.  If \code{year} cannot be coerced to an integer, the function will
throw an error}
}
\value{
This function returns a string of the form 'accident_YEAR.csv.bz2', with YEAR replaced
by whatever is passed by \code{year}.  A string is returned, although
a warning is givin if \code{year} cannot be coerced
to an integer, even though that may not form a valid filename.  No path is prepended
}
\description{
FARS files are of the form 'accident_YEAR.csv.bz2'.  This function takes \code{year}
as an input, and returns the file name in the correct form.  \code{year} must be a scalar
of a type coercible into an integer, otherwise it will throw an error.  No guarantee is made
that the file of the name created by this function exists
}
\examples{
make_filename(2013)
make_filename('2014')
\dontrun{fars_read(make_filename(2013))}
\dontrun{make_filename(12345)} # still returns a filename, but the file does not exist

}
\seealso{
\code{\link{fars_read}} To easily load the file named in the returned string
}

