% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate plot make of accidents in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric value for the specific state}

\item{year}{numeric value for a single year}
}
\value{
plot of the specified state with points representing the location of individual accidents
}
\description{
Generate plot make of accidents in a given state and year
}
\examples{
fars_map_state(1,2014)
}
