% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Fars map state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric. The US state to be plot.}

\item{year}{character. The year for which the function will plot the accident.}
}
\value{
A plot showing where the accident happened for a given state in a
given year.
}
\description{
This functions plots the fatal vehicles crashes location of a given year on a
state map of a given state. If no crashes happened in a given year/state, the
function returns a message saying that there is no accident to plot.
}
\examples{
fars_map_state(28,2013)

}
