% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number which represents a state}

\item{year}{string which defines a year, 2013, 2014, or 2015}
}
\value{
A plot shoing all traffic fatalities.
}
\description{
This function takes a state number and year, and returns a plot showing all traffic fatalities.
If the years are not 2013-2015, or the number doesn't correspond to a state, there will be an error.
}
