% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Open FARS dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of a comma-separated values text file}
}
\value{
table containing FARS data
}
\description{
Internal function that opens and parses a FARS dataset.
}
\examples{
fars_read('accident_2013.csv.bz2')
}
