% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character. Path to the csv file.}
}
\value{
tibble
}
\description{
This functions coerce data saved as csv to a tibble object.
The function returns a message if the file does not exist.
}
