% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file into a data frame table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file name that you want to read; it should be a character.}
}
\value{
This function returns a data frame table that stores data. However,
If the file you want to read does not exist, it gives you an error message.
}
\description{
This function reads data in a data file from the US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System.
}
\details{
\code{fars_read(filename)} utilizes \code{\link[readr]{read_csv}}.
In the end, this function provides a '\code{\link[dplyr]{tbl_df}}' class to a data frame.
See \code{\link[tibble]{tibble-package}} for more details.
}
\note{
If you need to read data for multiple years at a time, use \code{\link{fars_read_years}}.
In addition, you can use \code{\link{make_filename}} to generate a file name
more easily.
}
\examples{
\dontrun{make_filename(2013) \%>\% fars_read}

}

