% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load a file from the FARS database}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string containing the filename to be read from the FARS data.
be sure to include a valid path if the data are not in the current working directory.
See details below for finding the data}
}
\value{
This function returns a dplyr tibble containing the data.  If \code{filename} is not valid,
     the function stops with an error message
}
\description{
A function for loading the FARS data as specified by \code{filename}.
If \code{filename} exists it returns the data as a dplyr \code{\link{dplyr::tbl_df}}.  If it
does not exist, it stops with an error.
}
\details{
The FARS data can be found at the National Highway Traffic Safety Association
     \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
     If you choose to download it directly, be sure to include the correct path to the data.

Data for 2013-2015 are included in the distribution of this package in the \code{extdata}
     directory.  To find it, use the \code{system.file()} function.  For example,
     \code{system.file('extdata','accident_2013.csv.bz2',package = 'FARScoursera')}
}
\examples{
fars_read('accident_2013.csv.bz2')
fars_read(system.file('extdata','accident_2013.csv.bz2',package = 'FARScoursera'))
fars_read(make_filename(2013))
fars_read('not_a_valid_file') #shows the expected errors

}
\seealso{
\code{\link{make_filename}} for convenient creation of valid \code{filename} input
}

