% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string giving the file name of dataset}
}
\value{
A dataframe containing the corresponding data.
}
\description{
This function takes a string (file name), and returns a data frame containing the corresponding data.
If the filename is not correct it will results in an error.
}
