% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get accident data for each month of a given year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{numeric vector containing one or more years}
}
\value{
list containing table of month and year of accidents
}
\description{
Get accident data for each month of a given year
}
\examples{
fars_read_years(2014)

fars_read_years(c(2013, 2014, 2015))
}
