% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Fars read year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of numeric or character with the year the function will
return the data from}
}
\value{
A tibble by given year with 2 columns each for MONTH and year of the
event
}
\description{
This functions read the fars data from a given year and return a tibble with
month and year columns, one line by accident. The function returns an error if
the file for the input year is not found.
}
