% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing strings which contains 1 or more years.}
}
\value{
A list of dataframes
}
\description{
This function does the same as fars_read, but takes a vector of strings (years)
and returns a list containing dataframes.If the years are not 2013-2015, there will be an error.
}
