% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate monthly summary of accidents for a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{numeric vector containing one or more years}
}
\value{
table containing the number of accidents in each month during a given year
}
\description{
Generate monthly summary of accidents for a given year
}
\examples{
fars_summarize_years(c(2014, 2015))
}
