% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fars summarize years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of numeric or character with the year the function will
return the data from}
}
\value{
A tibble with number of monthly accidents. Months in rows and years in
columns.
}
\description{
This functions summarizes the monthly number of fatal accidents for the given
years.
}
\examples{
fars_summarize_years(c(2013,2014))
}
