% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing strings which contains 1 or more years.}
}
\value{
A tibble containing the number of fatal injuries by year and month
}
\description{
This function takes a vector of strings (years) and returns a tibble containing
the number of fatal injuries by year and month.If the years are not 2013-2015, there will be an error.
}
