% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric. Year of which the function will get the filename.}
}
\value{
A string as accident_XXXX.csv.bz2, XXXX being the input year,
ready to be use as a file name.
}
\description{
This functions insert a given year into the string accident_XXXX.csv.bz2.
}
