% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{string which defines a year, 2013, 2014, or 2015}
}
\value{
A file name (string) for the data corresponding to a particular year.
}
\description{
This function takes a string (year), and returns a string which is the file name of
the data for that particular year. If the year is not 2013-2015, there will be an error.
}
